test: Add swiftsync aggregate test
Introduce a trivial example of adding and subtracting elements from the
aggregate to catch mistakes in future modifications to the aggregate, if
required. Also proves basic facts about the aggregate: if the set of
elements added and the set of elements subtracted differ, then the
aggregate should be non-zero. If the sets agree, the internal state
should be zero.